Login, Logout, Restricted Pages and functionalities

+++ base/views.py
@@ -2,6 +2,12 @@ from django.shortcuts import render, redirect
 from .models import Room, Topic
 from .forms import RoomForm
 from django.db.models import Q
+from django.contrib.auth.models import User
+from django.contrib import messages
+from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.decorators import login_required
+from django.http import HttpResponse
+
 # Create your views here.
 
 # rooms = [
@@ -10,7 +16,32 @@ from django.db.models import Q
 #     {"id":3, "name":"Frontend developers"},
 # ]
 
+def loginPage(request):
+    if request.user.is_authenticated:
+        return redirect("home")
+    if request.method == "POST":
+        username = request.POST.get("username")
+        password = request.POST.get("password")
 
+        try:
+            user = User.objects.get(username=username)
+        except:
+            messages.error(request, "User does not exist.")
+
+        user = authenticate(request, username=username, password=password)
+
+        if user is not None:
+            login(request, user)    # creates the session inside the database and inside the browser
+            return redirect("home")
+        else:
+            messages.error(request, "Username OR Password is not correct.")
+    context = {}
+    return render(request, "base/login_register.html", context)
+
+
+def logoutUser(request):
+    logout(request)
+    return redirect("home")
 def home(request):
     q = request.GET.get('q') if request.GET.get('q') != None else ''
     rooms = Room.objects.filter(
@@ -28,6 +59,7 @@ def room(request, pk):
     context = {"room":room}
     return render(request, "base/room.html", context=context)
 
+@login_required(login_url='login')      # if the user is not logged in then we will redirect to login
 def createRoom(request):
     form = RoomForm
     if request.method == "POST":
@@ -39,9 +71,13 @@ def createRoom(request):
     context = {'form':form}
     return render(request, "base/room_form.html", context)
 
+@login_required(login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+    if request.user != room.host :
+        return HttpResponse("You are not allowed to update the room of someone else!!")
+    
     if request.method == "POST":
         form = RoomForm(request.POST, instance=room)
         if form.is_valid():
@@ -50,8 +86,12 @@ def updateRoom(request, pk):
     context = {'form':form}
     return render(request, "base/room_form.html", context)
 
+@login_required(login_url='login')
 def deleteRoom(request,pk):
     room = Room.objects.get(id=pk)
+    if request.user != room.host :
+        return HttpResponse("You are not allowed to update the room of someone else!!")
+    
     if request.method == "POST":
         room.delete()
         return redirect("home")
